requirements.txt
Flask==3.1.1
Werkzeug==3.1.0
Jinja2==3.1.6
python-dotenv==1.0.0
gunicorn==23.0.0
requests==2.32.4
